requirements/development.txt
-r common.txt

django-debug-toolbar==1.0.1
coverage
nose
django-nose